+++ lib/ui/profile/ProfileInfoSharePanel.dart
@@ -25,14 +25,18 @@ import 'package:sms_mms/sms_mms.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class ProfileInfoShareSheet extends StatelessWidget {
+  final Auth2UserProfile? publicProfile;
   final Auth2UserProfile? profile;
+  final Auth2UserPrivacy? privacy;
   final Uint8List? photoImageData;
   final Uint8List? pronunciationAudioData;
 
-  ProfileInfoShareSheet._({this.profile, this.photoImageData, this.pronunciationAudioData});
+  ProfileInfoShareSheet._({this.profile, this.privacy, this.photoImageData, this.pronunciationAudioData}) :
+    publicProfile = profile?.buildPublic(privacy, permitted: { Auth2FieldVisibility.public });
 
   static void present(BuildContext context, {
     Auth2UserProfile? profile,
+    Auth2UserPrivacy? privacy,
     Uint8List? photoImageData,
     Uint8List? pronunciationAudioData,
   }) {
@@ -48,6 +52,7 @@ class ProfileInfoShareSheet extends StatelessWidget {
       shape: RoundedRectangleBorder(borderRadius: BorderRadius.vertical(top: Radius.circular(16))),
       builder: (context) => ProfileInfoShareSheet._(
         profile: profile,
+        privacy: privacy,
         photoImageData: photoImageData,
         pronunciationAudioData: pronunciationAudioData,
       ),
@@ -57,7 +62,7 @@ class ProfileInfoShareSheet extends StatelessWidget {
   @override
   Widget build(BuildContext context) => Stack(children: [
       ProfileInfoShareWidget(
-        profile: profile,
+        publicProfile: publicProfile,
         photoImageData: photoImageData,
         pronunciationAudioData: pronunciationAudioData,
         topOffset: 24 + 2 * 16 /* close button size */,
@@ -86,16 +91,19 @@ class ProfileInfoSharePage extends StatefulWidget {
   static const String profileResultKey = 'profile_result';
 
   final Auth2UserProfile? _profile;
+  final Auth2UserPrivacy? _privacy;
   final Uint8List? _photoImageData;
   final Uint8List? _pronunciationAudioData;
   final Map<String, dynamic>? params;
 
-  ProfileInfoSharePage({Auth2UserProfile? profile, Uint8List? photoImageData, Uint8List? pronunciationAudioData, this.params, super.key}) :
+  ProfileInfoSharePage({Auth2UserProfile? profile, Auth2UserPrivacy? privacy, Uint8List? photoImageData, Uint8List? pronunciationAudioData, this.params, super.key}) :
     _profile = profile,
+    _privacy = privacy,
     _photoImageData = photoImageData,
     _pronunciationAudioData = pronunciationAudioData;
 
   Auth2UserProfile? get profile => _profile ?? _profileResultParam?.profile;
+  Auth2UserPrivacy? get privacy => _privacy ?? _profileResultParam?.privacy;
   Uint8List? get photoImageData => _photoImageData ?? _profileResultParam?.photoImageData;
   Uint8List? get pronunciationAudioData => _pronunciationAudioData ?? _profileResultParam?.pronunciationAudioData;
 
@@ -106,7 +114,9 @@ class ProfileInfoSharePage extends StatefulWidget {
 }
 
 class _ProfileInfoSharePageState extends State<ProfileInfoSharePage> {
+  Auth2UserProfile? _publicProfile;
   Auth2UserProfile? _profile;
+  Auth2UserPrivacy? _privacy;
   Uint8List? _photoImageData;
   Uint8List? _pronunciationAudioData;
   bool _loading = false;
@@ -122,7 +132,7 @@ class _ProfileInfoSharePageState extends State<ProfileInfoSharePage> {
     _loadingContent : _pageContent;
 
   Widget get _pageContent => ProfileInfoShareWidget(
-    profile: _profile,
+    publicProfile: _publicProfile,
     photoImageData: _photoImageData,
     pronunciationAudioData: _pronunciationAudioData
   );
@@ -136,12 +146,16 @@ class _ProfileInfoSharePageState extends State<ProfileInfoSharePage> {
   );
 
   Future<void> _loadInitialContent() async {
-    if (widget.profile != null) {
-      _profile = widget.profile;
-      _photoImageData = widget.photoImageData;
-      _pronunciationAudioData = widget.pronunciationAudioData;
-    }
-    else {
+    _profile = widget.profile;
+    _privacy = widget.privacy;
+    _photoImageData = widget.photoImageData;
+    _pronunciationAudioData = widget.pronunciationAudioData;
+    _publicProfile = _profile?.buildPublic(_privacy, permitted: { Auth2FieldVisibility.public });
+
+    if ((_profile == null) || (_privacy == null) ||
+        ((_profile?.photoUrl?.isNotEmpty == true) && (_photoImageData == null)) ||
+        ((_profile?.pronunciationUrl?.isNotEmpty == true) && (_pronunciationAudioData == null))
+    ) {
       setState(() {
         _loading = true;
       });
@@ -150,6 +164,7 @@ class _ProfileInfoSharePageState extends State<ProfileInfoSharePage> {
         _profile = loadResult.publicProfile();
         _photoImageData = loadResult.photoImageData;
         _pronunciationAudioData = loadResult.pronunciationAudioData;
+        _publicProfile = _profile?.buildPublic(_privacy, permitted: { Auth2FieldVisibility.public });
         _loading = false;
       });
     }
@@ -158,13 +173,13 @@ class _ProfileInfoSharePageState extends State<ProfileInfoSharePage> {
 
 class ProfileInfoShareWidget extends StatefulWidget {
 
-  final Auth2UserProfile? profile;
+  final Auth2UserProfile? publicProfile;
   final Uint8List? photoImageData;
   final Uint8List? pronunciationAudioData;
   final double topOffset;
   final double contentPaddingX;
 
-  ProfileInfoShareWidget({this.profile, this.photoImageData, this.pronunciationAudioData,
+  ProfileInfoShareWidget({this.publicProfile, this.photoImageData, this.pronunciationAudioData,
     this.topOffset = 16,
     this.contentPaddingX = 0,
   });
@@ -191,7 +206,7 @@ class _ProfileInfoShareWidgetState extends State<ProfileInfoShareWidget> {
       Column(children: [
         Padding(padding: EdgeInsets.symmetric(horizontal: widget.contentPaddingX), child:
           RepaintBoundary(key: _repaintBoundaryKey, child:
-            DirectoryAccountContactCard(account: Auth2PublicAccount(profile: widget.profile), printMode: true,),
+            DirectoryAccountContactCard(account: Auth2PublicAccount(profile: widget.publicProfile), printMode: true,),
           ),
         ),
         Padding(padding: EdgeInsets.symmetric(vertical: 16), child:
@@ -205,7 +220,7 @@ class _ProfileInfoShareWidgetState extends State<ProfileInfoShareWidget> {
         )),
         _buildCommand(
           icon: Styles().images.getImage('envelope', size: _commandIconSize),
-          text: Localization().getStringEx('panel.profile.info.share.command.button.share.email.text', 'Share Digital Business Card'),
+          text: Localization().getStringEx('panel.profile.info.share.command.button.share.email.text', 'Share via Email'),
           progress: _preparingEmail,
           onTap: _onTapShareViaEmail,
         ),
@@ -217,7 +232,7 @@ class _ProfileInfoShareWidgetState extends State<ProfileInfoShareWidget> {
         ),
         _buildCommand(
           icon: Styles().images.getImage('up-from-bracket', size: _commandIconSize),
-          text: Localization().getStringEx('panel.profile.info.share.command.button.share.digital_card.text', 'Add to Device Contacts'),
+          text: Localization().getStringEx('panel.profile.info.share.command.button.share.digital_card.text', 'Share My QR Code'),
           progress: _sharingDigitalCard,
           onTap: _onTapShareDigitalCard,
         ),
@@ -274,7 +289,7 @@ class _ProfileInfoShareWidgetState extends State<ProfileInfoShareWidget> {
   void _onTapShareDigitalCard() async {
     Analytics().logSelect(target: 'Share Digital Card');
     QrCodePanel.presentProfile(context,
-      profile: widget.profile,
+      profile: widget.publicProfile,
       photoImageData: widget.photoImageData,
       pronunciationAudioData: widget.pronunciationAudioData,
     );
@@ -328,8 +343,7 @@ class _ProfileInfoShareWidgetState extends State<ProfileInfoShareWidget> {
         _preparingEmail = false;
       });
 
-      String? emailBody = widget.profile?.toDisplayText() ?? '';
-
+      String? emailBody = widget.publicProfile?.toDisplayText() ?? '';
       if (kIsWeb) {
         final Map<String, String> parameters = <String, String>{
           'body': emailBody
@@ -373,7 +387,7 @@ class _ProfileInfoShareWidgetState extends State<ProfileInfoShareWidget> {
       setState(() {
         _preparingTextMessage = false;
       });
-      String smsBody = widget.profile?.toDisplayText() ?? '';
+      String smsBody = widget.publicProfile?.toDisplayText() ?? '';
       if (kIsWeb) {
         final Map<String, String> parameters = <String, String>{
           'body': smsBody
@@ -437,7 +451,7 @@ class _ProfileInfoShareWidgetState extends State<ProfileInfoShareWidget> {
   Future<String?> _saveImage({bool addToGallery = false} ) async {
     try {
       double pixelRatio = MediaQuery.of(context).devicePixelRatio * 3;
-      final String saveFileName = '${widget.profile?.vcardFullName} ${DateTimeUtils.localDateTimeFileStampToString(DateTime.now())}.png';
+      final String saveFileName = '${widget.publicProfile?.vcardFullName} ${DateTimeUtils.localDateTimeFileStampToString(DateTime.now())}.png';
       if (kIsWeb) {
         //TBD: DD - implement
         AppAlert.showDialogResult(context, 'Not supported in web.');
@@ -470,12 +484,12 @@ class _ProfileInfoShareWidgetState extends State<ProfileInfoShareWidget> {
 
   Future<String?> _saveDigitalCard() async {
     try {
-      String? vcfContent = widget.profile?.toDigitalCard(
+      String? vcfContent = widget.publicProfile?.toDigitalCard(
         photoImageData: widget.photoImageData,
       );
       if ((vcfContent != null) && vcfContent.isNotEmpty) {
         final String dir = (await getApplicationDocumentsDirectory()).path;
-        final String saveFileName = '${widget.profile?.vcardFullName} ${DateTimeUtils.localDateTimeFileStampToString(DateTime.now())}';
+        final String saveFileName = '${widget.publicProfile?.vcardFullName} ${DateTimeUtils.localDateTimeFileStampToString(DateTime.now())}';
         final String fullPath = '$dir/$saveFileName.vcf';
         File capturedFile = File(fullPath);
         await capturedFile.writeAsString(vcfContent);
@@ -488,7 +502,7 @@ class _ProfileInfoShareWidgetState extends State<ProfileInfoShareWidget> {
 
   Future<bool> _copyTextToClipbiard() async {
     try {
-      await Clipboard.setData(ClipboardData(text: widget.profile?.toDisplayText() ?? ''));
+      await Clipboard.setData(ClipboardData(text: widget.publicProfile?.toDisplayText() ?? ''));
       return true;
     }
     catch(e) {